fix: Create new merged patch by renaming `Hide player buttons`
