+++ src/main/resources/addresources/values/strings.xml
@@ -518,7 +518,7 @@ This is because Crowdin requires temporarily flattening this file and removing t
             <string name="revanced_hide_player_flyout_watch_in_vr_summary_on">Watch in VR menu is hidden</string>
             <string name="revanced_hide_player_flyout_watch_in_vr_summary_off">Watch in VR menu is shown</string>
         </patch>
-        <patch id="layout.buttons.player.hide.HidePlayerButtonsPatch">
+        <patch id="layout.buttons.player.hide.HidePlayerOverlayButtonsPatch">
             <string name="revanced_hide_player_buttons_title">Hide previous &amp; next video buttons</string>
             <string name="revanced_hide_player_buttons_summary_on">Buttons are hidden</string>
             <string name="revanced_hide_player_buttons_summary_off">Buttons are shown</string>
